+++ src/components/AssetList/AssetListItem.tsx
@@ -1,16 +1,17 @@
 import { observer } from 'mobx-react-lite';
-import { ListItem, ListItemIcon, ListItemText, ListItemProps } from '@cere-wallet/ui';
+import { ListItem, ListItemIcon, ListItemText, ListItemProps, IconProps } from '@cere-wallet/ui';
 import { Asset } from '~/stores';
 import { CoinIcon } from '../CoinIcon';
 import { useBalanceStore } from '~/hooks';
 
-export type AssetListItemProps = ListItemProps & {
+export type AssetListItemProps = ListItemProps<'div'> & {
   asset: Asset;
+  disableInset?: boolean;
+  iconSize?: IconProps['fontSize'];
 };
 
-export const AssetListItem = ({ asset, ...props }: AssetListItemProps) => {
+export const AssetListItem = ({ asset, disableInset = false, iconSize = 'inherit', ...props }: AssetListItemProps) => {
   const { ticker, displayName, network, balance } = asset;
-
   const { getUsdBalance } = useBalanceStore();
 
   return (