+++ src/components/AssetList/index.ts
@@ -1,1 +1,2 @@
 export { default as AssetList, type AssetListProps } from './AssetList';
+export { default as AssetListItem, type AssetListItemProps } from './AssetListItem';