+++ src/components/Transfer/index.ts
@@ -1,0 +1,2 @@
+export * from './TransferFaq';
+export * from './TransferAsset';