+++ src/stores/ApprovalStore/ApprovalStore.ts
@@ -16,6 +16,10 @@ import { TransactionPopupState } from '../TransactionPopupStore';
 import { ConfirmPopupState } from '../ConfirmPopupStore';
 import { AppContextStore } from '../AppContextStore';
 
+type ApproveTransactionOptions = {
+  showDetails?: boolean;
+};
+
 const convertPrice = (amount: BigNumber, { decimals }: TokenConfig) => {
   return amount.div(10 ** decimals).toNumber();
 };
@@ -32,8 +36,8 @@ export class ApprovalStore {
 
   async approvePersonalSign({ preopenInstanceId, params: [content] }: PersonalSignRequest) {
     const tokenConfig = getTokenConfig();
-
-    const popup = await this.popupManagerStore.proceedTo<ConfirmPopupState>(preopenInstanceId, '/confirm', {
+    const instanceId = preopenInstanceId || this.popupManagerStore.createModal();
+    const popup = await this.popupManagerStore.proceedTo<ConfirmPopupState>(instanceId, '/confirm', {
       network: this.networkStore.network,
       app: this.contextStore.app,
       status: 'pending',
@@ -42,7 +46,7 @@ export class ApprovalStore {
     });
 
     await Promise.race([when(() => !popup.isConnected), when(() => popup.state.status !== 'pending')]);
-    this.popupManagerStore.closePopup(preopenInstanceId);
+    this.popupManagerStore.closePopup(instanceId);
 
     if (!popup.isConnected) {
       throw new Error('User has closed the confirmation popup');
@@ -53,14 +57,19 @@ export class ApprovalStore {
     }
   }
 
-  async approveSendTransaction({ preopenInstanceId, params: [transaction] }: SendTransactionRequest) {
+  async approveSendTransaction(
+    { preopenInstanceId, proceed, params: [transaction] }: SendTransactionRequest,
+    { showDetails = false }: ApproveTransactionOptions = {},
+  ) {
     const tokenConfig = getTokenConfig();
     const network = this.networkStore.network!;
     const { contractName, description: parsedData } = parseTransactionData(transaction, network.chainId);
-
-    const popup = await this.popupManagerStore.proceedTo<TransactionPopupState>(preopenInstanceId, '/transaction', {
+    const instanceId = preopenInstanceId || this.popupManagerStore.createModal();
+    const popup = await this.popupManagerStore.proceedTo<TransactionPopupState>(instanceId, '/transaction', {
       network,
       parsedData,
+      status: 'pending',
+      step: 'confirmation',
       from: transaction.from,
       to: transaction.to,
       rawData: transaction.data,
@@ -111,19 +120,52 @@ export class ApprovalStore {
           };
         }
       }
-
-      popup.state.status = 'pending';
     });
 
     await Promise.race([when(() => !popup.isConnected), when(() => popup.state.status !== 'pending')]);
-    this.popupManagerStore.closePopup(preopenInstanceId);
 
-    if (!popup.isConnected) {
-      throw new Error('User has closed the confirmation popup');
+    if (!popup.isConnected || popup.state.status === 'declined') {
+      this.popupManagerStore.closePopup(instanceId);
+
+      throw new Error(
+        popup.isConnected ? 'User has declined the transaction request' : 'User has closed the confirmation popup',
+      );
     }
 
-    if (popup.state.status === 'declined') {
-      throw new Error('User has declined the transaction request');
+    if (!showDetails) {
+      return this.popupManagerStore.closePopup(instanceId);
     }
+
+    when(
+      () => !popup.isConnected || popup.state.status === 'done',
+      () => {
+        this.popupManagerStore.closePopup(instanceId);
+      },
+    );
+
+    let isRejected = false;
+
+    try {
+      const { result: transactionId } = await proceed();
+
+      runInAction(() => {
+        popup.state.step = 'details';
+        popup.state.transaction = {
+          id: transactionId!,
+          status: 'pending',
+        };
+      });
+
+      const pendingTx = await this.wallet.provider!.getTransaction(transactionId!);
+      const txReceipt = await pendingTx.wait();
+
+      isRejected = !txReceipt.status;
+    } catch {
+      isRejected = true;
+    }
+
+    runInAction(() => {
+      popup.state.transaction!.status = isRejected ? 'rejected' : 'confirmed';
+    });
   }
 }