+++ CHANGELOG.md
@@ -1,8 +1,15 @@
 ## Release Notes:
 
 ### vNext
-- [DAV-3719] Fix console error on snackbar apperance 
 - [DAV-3491] Add custom asset management
+- [DAV-3719] Fix snackbar tour error in console
+- [DAV-3792] Align asset transfer UI with the design
+- [DAV-3734] Add in-wallet transaction modals
+- [DAV-3669] Add asset transfer (ethereum)
+
+### v1.12.0
+- [DAV-3702] Fix start point for product tour
+- [DAV-3767] Add GTM integration and click IDs
 
 ### v.1.11.0